Splitting main javascript into smaller chunks
