+++ scripts/config/logProvider.js
@@ -1,0 +1,4 @@
+angular.module('quizApp')
+    .config(['$logProvider', function($logProvider){
+        $logProvider.debugEnabled(true);
+    }]);
