+++ scripts/config/settings.js
@@ -1,0 +1,8 @@
+var settings = {
+    maxTime: 20000, //max time in ms
+    maxScore: 200,
+    minScore: 10,
+    gracePeriod: 2000
+}
+
+module.exports = settings;
