+++ scripts/controllers/QuizzesController.js
@@ -1,0 +1,52 @@
+angular.module('quizApp')
+    .controller('QuizzesController', ['QuizData', '$log', '$location', '$timeout', function(QuizData, $log, $location, $timeout){
+    var self = this;
+
+    self.loading = true;
+
+    self.startQuiz = function(categoryId,quizId){
+        $log.debug("Selected Quiz: ", categoryId,quizId);
+        QuizData.selectQuiz(categoryId,quizId,function() {
+            $location.path("/quiz/intro")
+        });
+    };
+
+    self.reloadQuizzes = function(){
+        $log.debug("Reloading Quizzes");
+        self.loading = true;
+
+        $timeout(function(){
+            self.reloadActive = true;
+        }, 2000);
+
+        QuizData.login(null,function(err, res){
+            if(!err){
+                self.quizzes = QuizData.getQuizzes();
+                self.categories = QuizData.getCategories();
+                self.loading = false;
+            }else{
+                $log.error("Unable to refresh quizzes", err);
+            }
+       });
+    };
+
+    self.logout = function(){
+        QuizData.logout();
+    };
+
+    self.studentData = QuizData.getStudentData();
+    if (self.studentData.classCode==undefined) {
+        self.categories = QuizData.getCategories();
+    }
+    else {
+        QuizData.login(null,function(err, res){
+            if(!err){
+                self.quizzes = QuizData.getQuizzes();
+                self.categories = QuizData.getCategories();
+                self.loading = false;
+            }else{
+                $log.error("Unable to refresh quizzes", err);
+            }
+       });
+    }
+}]);
