+++ views/index.jade
@@ -19,14 +19,11 @@ html(ng-app='quizApp')
             var localStateSet = #{isLocal};
             var wso2 = "#{wso2Enabled}";
 
-        script(src='/js/jquery-1.11.0.js')
-        script(src='/js/uuid.js')
-        script(src='/js/fastclick.js')
+        script(src='/js/vendor.js')
+        script(src='/lib/jquery-1.11.0.js')
+        script(src='/lib/uuid.js')
         script(src='#{session.zzishsdkurl}/zzish.js')
-        script(src='/js/bootstrap.js')
-        script(src='/js/angular.min.js')
-        script(src='/js/angular-route.min.js')
-        script(src='/js/angular-animate.min.js')
+        script(src='/lib/bootstrap.js')
         script(src='/js/quizApp.js')
         script(src="/outdatedbrowser/outdatedbrowser.min.js")
         script.
